Added e-mail function for sending e-mails #2
The function has void return, takes four parameters; from (the e-mail address of the sender), to (the e-mail address of the recipient), subject (the title/subject of the e-mail) and text (the content of the e-mail). These might be subject to changes, as I have not figured out what host to use for the e-mails yet.