Add javadoc task in Ant build file
Execute "ant doc" to generate the doc from both the main code and the tests.
Produce doc is in html and is stored in the doc directory.
This is part of the solution of #15 (it remains to generate the doc and push it to git in end of the project)